plagiarism_turnitin: Stop invalidating submissions that have a report
The cron task to check for submission updates invalidates the submission
if the corresponding paper ID cannot be found via the Turnitin API.
For BJTU this was causing all valid submissions to be invalidated
whenever the connection was unavailable. In the case of submissions
that already have a similarity report this is obviously not correct.